Reorganiza componentes para visualização das casas
