+++ client/src/app/aderencia/aderencia-parlamentares/aderencia-parlamentares.component.html
@@ -1,0 +1,112 @@
+<div class="container pt-3 container-sticky-footer-navbar">
+
+  <pagination-template
+    #pag="paginationApi"
+    maxSize="5"
+    (pageChange)="pageChange($event)"
+    [hidden]="pag.isFirstPage()"
+  >
+    <nav
+      aria-label="Paginação"
+      class="pb-1"
+    >
+      <ul class="pagination justify-content-center">
+        <li class="page-item">
+          <button
+            type="button"
+            class="page-link"
+            [disabled]="pag.isFirstPage()"
+            (click)="pag.previous()"
+          >
+            <span class="icon-chevron-left"></span>
+          </button>
+        </li>
+        <li
+          class="page-item"
+          *ngFor="let page of pag.pages"
+          [class.active]="pag.getCurrent() === page.value"
+        >
+          <button
+            type="button"
+            class="page-link"
+            (click)="pag.setCurrent(page.value)"
+          >
+            {{ page.label }}
+          </button>
+        </li>
+        <li class="page-item">
+          <button
+            type="button"
+            class="page-link"
+            [disabled]="pag.isLastPage()"
+            (click)="pag.next()"
+          >
+            <span class="icon-chevron-right"></span>
+          </button>
+        </li>
+      </ul>
+    </nav>
+  </pagination-template>
+
+  <div class="parlamentar-lista py-3">
+    <app-card-parlamentar
+      *ngFor="let parlamentar of parlamentares | paginate: { itemsPerPage: 20, currentPage: p }; index as i;"
+      [parlamentar]="parlamentar"
+      [id]="getParlamentarPosition(i, 20, p)"
+      [tema]="filtro?.tema"
+      [temaSlug]="filtro?.temaSlug"
+      [comissao]="filtro?.comissao"
+      [lideranca]="filtro?.lideranca"
+      [cargoComissao]="filtro?.cargoComissao"
+      [view]="view"
+      (followChecked)="onFollow($event)"
+    ></app-card-parlamentar>
+  </div>
+
+  <pagination-template
+    #pag="paginationApi"
+    maxSize="5"
+    (pageChange)="pageChange($event)"
+  >
+    <nav
+      aria-label="Paginação"
+      class="pb-1"
+    >
+      <ul class="pagination justify-content-center">
+        <li class="page-item">
+          <button
+            type="button"
+            class="page-link"
+            [disabled]="pag.isFirstPage()"
+            (click)="pag.previous()"
+          >
+            <span class="icon-chevron-left"></span>
+          </button>
+        </li>
+        <li
+          class="page-item"
+          *ngFor="let page of pag.pages"
+          [class.active]="pag.getCurrent() === page.value"
+        >
+          <button
+            type="button"
+            class="page-link"
+            (click)="pag.setCurrent(page.value)"
+          >
+            {{ page.label }}
+          </button>
+        </li>
+        <li class="page-item">
+          <button
+            type="button"
+            class="page-link"
+            [disabled]="pag.isLastPage()"
+            (click)="pag.next()"
+          >
+            <span class="icon-chevron-right"></span>
+          </button>
+        </li>
+      </ul>
+    </nav>
+  </pagination-template>
+</div>
