+++ client/src/app/aderencia/aderencia.component.html
@@ -1,6 +1,7 @@
 <app-sticky-footer-navbar></app-sticky-footer-navbar>
-<div class="container pt-3 container-sticky-footer-navbar">
-  <h4 class="title-main">Alinhamento dos deputados ao Governo</h4>
+<div class="container py-4 container-sticky-footer-navbar">
+  <h4 class="title-main">Alinhamento da {{ casa }} ao Governo</h4>
+
   <app-filter (filterChange)="search($event)"></app-filter>
 
   <header class="controls row no-gutters my-3">
@@ -11,6 +12,18 @@
     <div class="col-8 col-md-12">
       <div class="btn-toolbar">
         <div class="d-none d-md-block">Exibir:</div>
+        <button
+          [ngClass]="{'btn btn-light btn-light-icon': true, 'active': view === VIEW_ARC}"
+          (click)="setView('arc')"
+        >
+          <span class="icon-arc"></span>
+        </button>
+        <button
+          [ngClass]="{'btn btn-light btn-light-icon': true, 'active': view === VIEW_BEE}"
+          (click)="setView('bee')"
+        >
+          <span class="icon-beeswarm"></span>
+        </button>
         <button
           [ngClass]="{'btn btn-light btn-light-icon': true, 'active': view == VIEW_LG}"
           (click)="setView('lg')"
@@ -40,113 +53,26 @@
     </div>
   </header>
 
-  <pagination-template
-    #pag="paginationApi"
-    maxSize="5"
-    (pageChange)="pageChange($event)"
-    [hidden]="pag.isFirstPage()"
-  >
-    <nav
-      aria-label="Paginação"
-      class="pb-1"
-    >
-      <ul class="pagination justify-content-center">
-        <li class="page-item">
-          <button
-            type="button"
-            class="page-link"
-            [disabled]="pag.isFirstPage()"
-            (click)="pag.previous()"
-          >
-            <span class="icon-chevron-left"></span>
-          </button>
-        </li>
-        <li
-          class="page-item"
-          *ngFor="let page of pag.pages"
-          [class.active]="pag.getCurrent() === page.value"
-        >
-          <button
-            type="button"
-            class="page-link"
-            (click)="pag.setCurrent(page.value)"
-          >
-            {{ page.label }}
-          </button>
-        </li>
-        <li class="page-item">
-          <button
-            type="button"
-            class="page-link"
-            [disabled]="pag.isLastPage()"
-            (click)="pag.next()"
-          >
-            <span class="icon-chevron-right"></span>
-          </button>
-        </li>
-      </ul>
-    </nav>
-  </pagination-template>
-
-  <div class="parlamentar-lista py-3">
-    <app-card-parlamentar
-      *ngFor="let parlamentar of parlamentares | paginate: { itemsPerPage: 20, currentPage: p }; index as i;"
-      [parlamentar]="parlamentar"
-      [id]="getParlamentarPosition(i, 20, p)"
-      [tema]="filtro?.tema"
-      [temaSlug]="filtro?.temaSlug"
-      [comissao]="filtro?.comissao"
-      [lideranca]="filtro?.lideranca"
-      [cargoComissao]="filtro?.cargoComissao"
+  <div [hidden]="!getVisaoPlenario()">
+    <app-congresso-chart
+      [parlamentares]="parlamentares"
+      [parlamentaresCompleto]="parlamentaresCompleto"
       [view]="view"
-      (followChecked)="onFollow($event)"
-    ></app-card-parlamentar>
+      [filter]="filtro"
+      [modo]="'aderencia'"
+      class="plot"
+      (viewEvent)="setView($event)"
+      (finishEvent)="setChartVisibility($event)"
+    ></app-congresso-chart>
+
+    <app-congresso-chart-legenda></app-congresso-chart-legenda>
+  </div>
+  <div [hidden]="getVisaoPlenario()">
+    <app-aderencia-parlamentares
+      [parlamentares]="parlamentares"
+      [view]="view"
+      [orderBy]="orderBy"
+    ></app-aderencia-parlamentares>
   </div>
 
-  <pagination-template
-    #pag="paginationApi"
-    maxSize="5"
-    (pageChange)="pageChange($event)"
-  >
-    <nav
-      aria-label="Paginação"
-      class="pb-1"
-    >
-      <ul class="pagination justify-content-center">
-        <li class="page-item">
-          <button
-            type="button"
-            class="page-link"
-            [disabled]="pag.isFirstPage()"
-            (click)="pag.previous()"
-          >
-            <span class="icon-chevron-left"></span>
-          </button>
-        </li>
-        <li
-          class="page-item"
-          *ngFor="let page of pag.pages"
-          [class.active]="pag.getCurrent() === page.value"
-        >
-          <button
-            type="button"
-            class="page-link"
-            (click)="pag.setCurrent(page.value)"
-          >
-            {{ page.label }}
-          </button>
-        </li>
-        <li class="page-item">
-          <button
-            type="button"
-            class="page-link"
-            [disabled]="pag.isLastPage()"
-            (click)="pag.next()"
-          >
-            <span class="icon-chevron-right"></span>
-          </button>
-        </li>
-      </ul>
-    </nav>
-  </pagination-template>
 </div>
