+++ client/src/app/aderencia/aderencia.component.scss
@@ -1,35 +1,36 @@
-.parlamentar-count {
-  line-height: 3;
-  font-size: 0.8rem;
-
-  @media (min-width: 992px) {
-    line-height: 2;
-    font-size: 1rem;
-  }
+.subtitle-wrapper {
+  height: 50px;
 }
 
-.parlamentar-lista {
-  @media (min-width: 992px) {
-    width: 760px;
-    margin: auto;
-  }
+.subtitle {
+  flex-wrap: wrap;
 }
 
 .btn-toolbar {
-  line-height: 2.2rem;
+  line-height: 3.2rem;
   justify-content: flex-end;
 
   @media (min-width: 992px) {
     justify-content: center;
   }
+
 }
 
-.controls {
-  @media (max-width: 575.98px) {
-    position: -webkit-sticky;
-    position: sticky;
-    top: 93px;
-    background-color: #ffffff;
-    z-index: 1050;
+.chart-controls-info {
+  order: 0;
+  line-height: 3.2rem;
+}
+
+.chart-controls-info2 {
+  order: 2;
+  @media (min-width: 992px) {
+    order: 1;
+  }
+}
+
+.chart-controls-btns {
+  order: 1;
+  @media (min-width: 992px) {
+    order: 2;
   }
 }
