+++ client/src/app/aderencia/aderencia.module.ts
@@ -10,12 +10,18 @@ import { AderenciaRoutingModule } from './aderencia-routing.module';
 import { FilterModule } from '../filter/filter.module';
 
 import { AderenciaComponent } from './aderencia.component';
+import { AderenciaParlamentaresComponent } from './aderencia-parlamentares/aderencia-parlamentares.component';
 import { CardParlamentarComponent } from './card-parlamentar/card-parlamentar.component';
+import { CongressoChartLegendaComponent } from './congresso-chart-legenda/congresso-chart-legenda.component';
+import { CongressoChartComponent } from './congresso-chart/congresso-chart.component';
 
 @NgModule({
   declarations: [
     AderenciaComponent,
+    AderenciaParlamentaresComponent,
     CardParlamentarComponent,
+    CongressoChartLegendaComponent,
+    CongressoChartComponent
   ],
   imports: [
     CommonModule,
