+++ client/src/app/aderencia/congresso-chart-legenda/congresso-chart-legenda.component.html
@@ -1,0 +1,52 @@
+<div class="subtitle-wrapper">
+  <div class="d-flex justify-content-center subtitle">
+    <div>
+      <app-legend
+        [title]="'100-80%'"
+        [class]="'alinhamento-5'"
+        [striped]="false"
+        [rounded]="true"
+      ></app-legend>
+    </div>
+    <div>
+      <app-legend
+        [title]="'80-60%'"
+        [class]="'alinhamento-4'"
+        [striped]="false"
+        [rounded]="true"
+      ></app-legend>
+    </div>
+    <div>
+      <app-legend
+        [title]="'60-50%'"
+        [class]="'alinhamento-3'"
+        [striped]="false"
+        [rounded]="true"
+      ></app-legend>
+    </div>
+    <div>
+      <app-legend
+        [title]="'50-30%'"
+        [class]="'alinhamento-2'"
+        [striped]="false"
+        [rounded]="true"
+      ></app-legend>
+    </div>
+    <div>
+      <app-legend
+        [title]="'30-0%'"
+        [class]="'alinhamento-1'"
+        [striped]="false"
+        [rounded]="true"
+      ></app-legend>
+    </div>
+    <div>
+      <app-legend
+        [title]="'não calculado'"
+        [class]="'alinhamento-indefinido'"
+        [striped]="false"
+        [rounded]="true"
+      ></app-legend>
+    </div>
+  </div>
+</div>
