+++ client/src/app/aderencia/congresso-chart-legenda/congresso-chart-legenda.component.scss
@@ -1,0 +1,7 @@
+.subtitle-wrapper {
+  height: 50px;
+}
+
+.subtitle {
+  flex-wrap: wrap;
+}
