+++ client/src/app/aderencia/congresso-chart-legenda/congresso-chart-legenda.component.ts
@@ -1,0 +1,12 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-congresso-chart-legenda',
+  templateUrl: './congresso-chart-legenda.component.html',
+  styleUrls: ['./congresso-chart-legenda.component.scss']
+})
+export class CongressoChartLegendaComponent {
+
+  constructor() { }
+
+}
