+++ client/src/app/aderencia/congresso-chart/congresso-chart.component.scss
