+++ client/src/app/app-routing.module.ts
@@ -4,10 +4,10 @@ import { Routes, RouterModule } from '@angular/router';
 const routes: Routes = [
   { path: '', loadChildren: './main/main.module#MainModule' },
   { path: 'usuario', loadChildren: './user/user.module#UserModule' },
-  { path: 'questionario', loadChildren: './questionario/questionario.module#QuestionarioModule' },
-  { path: 'alinhamento', loadChildren: './alinhamento/alinhamento.module#AlinhamentoModule' },
+  // { path: 'questionario', loadChildren: './questionario/questionario.module#QuestionarioModule' },
+  // { path: 'alinhamento', loadChildren: './alinhamento/alinhamento.module#AlinhamentoModule' },
   { path: 'parlamentar', loadChildren: './parlamentar/parlamentar.module#ParlamentarModule' },
-  { path: 'congresso', loadChildren: './congresso/congresso.module#CongressoModule' },
+  // { path: 'congresso', loadChildren: './congresso/congresso.module#CongressoModule' },
   { path: 'aderencia', loadChildren: './aderencia/aderencia.module#AderenciaModule' },
   { path: '**', redirectTo: '' }
 ];
