+++ client/src/app/shared/components/navbar/navbar.component.html
@@ -17,7 +17,7 @@
       <ul class="nav nav-pills nav-fill">
         <li class="nav-item">
           <a
-            [routerLink]="['/congresso/aderencia/camara']"
+            [routerLink]="['/aderencia/camara']"
             [routerLinkActive]="'active'"
             class="nav-link navbar-nav-va-link"
             href="#"
@@ -27,7 +27,7 @@
         </li>
         <li class="nav-item">
           <a
-            [routerLink]="['/congresso/aderencia/senado']"
+            [routerLink]="['/aderencia/senado']"
             [routerLinkActive]="'active'"
             class="nav-link navbar-nav-va-link"
             href="#"
